packages/tools/forensics/ussy-strata/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ussy-strata"
version = "0.1.0"
description = "Git forensics with geological metaphors — merged stratagit + unconformity"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["git", "forensics", "geological", "stratigraphy", "visualization"]
dependencies = [
    "click>=8.1",
    "gitpython>=3.1",
    "rich>=13.0",
    "rich-click>=1.0",
    "ussy-core",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
ussy-strata = "ussy_strata.cli:main"
stratagit = "ussy_strata.legacy:stratagit"
unconformity = "ussy_strata.legacy:unconformity"
disconformitussy = "ussy_strata.legacy:disconformitussy"

[tool.hatch.build.targets.wheel]
packages = ["src/ussy_strata"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
